Refactor QR Scanning component usage
- Replace `QRCodeThumbnail` with `CoveredQRCodeScanner` in `Send`.
- Remove `QRCodeThumbnail`, as that was its only place of usage.
